Label: Use assignedDate on the household_size label.
Fix: #307
