+++ survey/locales/fr/home.yaml
@@ -28,7 +28,8 @@ home.geography: >-
     Une fois le lieu localisé, vous pouvez déplacer le point pour plus de précision.
 household.size: >-
     <strong>Personnes habitant votre domicile</strong> ou logement.<br /><span class="_pale
-    _oblique">Y compris vous-même.</span>
+    _oblique">Y compris vous-même et les enfants en garde partagée s'ils étaient présents
+    le {{assignedDate}}.</span>
 household.carNumber: >-
     <strong>Automobiles</strong>, VUS, camions légers ou fourgonnettes, etc.<br /><span
     class="_pale _oblique">À la disposition de votre ménage.</span>
