+++ survey/references/questionnaire_dictionary_en.csv
@@ -54,9 +54,9 @@ Question,"Positioning your home
 Search for your home using its postal code or address by clicking the button below, or please position the location on the map by navigating, zooming, and clicking. Once the location is located, you can move the point for greater precision."
 ""
 Field,hh_size
-Question type,RadioNumber
+Question type,Unknown input
 Question,"Persons living in your home
-Including yourself."
+Including yourself and children in shared custody if they were present on {{assignedDate}}."
 ""
 Field,hh_carNumber
 Question type,RadioNumber
