+++ survey/references/questionnaire_dictionary_fr.csv
@@ -54,9 +54,9 @@ Question,"Positionnement de votre domicile
 Recherchez votre domicile en utilisant son code postal ou son adresse en cliquant sur le bouton ci-dessous, ou veuillez positionner le lieu sur la carte en naviguant, zoomant et cliquant. Une fois le lieu localisé, vous pouvez déplacer le point pour plus de précision."
 ""
 Champ,hh_size
-Type de question,RadioNumber
+Type de question,Entrée inconnue
 Question,"Personnes habitant votre domicile ou logement.
-Y compris vous-même."
+Y compris vous-même et les enfants en garde partagée s'ils étaient présents le {{assignedDate}}."
 ""
 Champ,hh_carNumber
 Type de question,RadioNumber
