+++ survey/references/questionnaire_list_fr.txt
@@ -29,7 +29,7 @@ Positionnement de votre domicile
 Recherchez votre domicile en utilisant son code postal ou son adresse en cliquant sur le bouton ci-dessous, ou veuillez positionner le lieu sur la carte en naviguant, zoomant et cliquant. Une fois le lieu localisé, vous pouvez déplacer le point pour plus de précision.
 
 Personnes habitant votre domicile ou logement.
-Y compris vous-même.
+Y compris vous-même et les enfants en garde partagée s'ils étaient présents le {{assignedDate}}.
 
 Automobiles, VUS, camions légers ou fourgonnettes, etc.
 À la disposition de votre ménage.
